Phase 6 canonical cleanup for webhook and portal request paths
