+++ app/api/settings/update/shop/owner-pin/set/route.ts
@@ -1,1 +1,3 @@
+// Backward-compatible alias for older settings clients.
+// Canonical endpoint is /api/shop/owner-pin/set.
 export { POST } from "../../../../../shop/owner-pin/set/route";
