+++ app/api/stripe/checkout/webhook/route.ts
@@ -1,6 +1,6 @@
 export const runtime = "nodejs";
 
-import { handleStripeWebhook } from "@/features/stripe/api/stripe/checkout/webhook/route";
+import { handleStripeWebhook } from "@/features/stripe/api/stripe/webhook/route";
 
 // Backward-compatible alias for legacy endpoint configuration.
 // Canonical endpoint is /api/stripe/webhook.
