+++ app/api/stripe/webhook/route.ts
@@ -1,6 +1,6 @@
 export const runtime = "nodejs";
 
-import { handleStripeWebhook } from "@/features/stripe/api/stripe/checkout/webhook/route";
+import { handleStripeWebhook } from "@/features/stripe/api/stripe/webhook/route";
 
 export async function POST(req: Request) {
   return handleStripeWebhook(req);
