+++ features/stripe/api/stripe/webhook/route.ts
@@ -1,0 +1,1 @@
+export { runtime, handleStripeWebhook } from "@/features/stripe/api/stripe/checkout/webhook/route";
